scripts/requirements.txt
psycopg2-binary==2.8.6
requests==2.25.1
pandas==1.2.4
matplotlib==3.4.2
schedule==1.1.0